first student question some changes

+++ app/src/main/java/com/example/simplekahoot/Student.kt
@@ -1,4 +1,4 @@
 package com.example.simplekahoot
 
-class Student(var StudentName:String, var Score:Int=0) {
+class Student(var StudentName:String, var Score:Double=0.0) {
 }
